% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auscophub_functions.R
\name{s_mk_fold}
\alias{s_mk_fold}
\title{Internal function for constructing archive folders for Sentinel-2 data.}
\usage{
s_mk_fold(dzip = dzip, path = path)
}
\arguments{
\item{dzip}{is a character filepath for one zipped Sentinel download which
is passed from the parent function.}

\item{path}{is a character filepath to the location of the Sentinel zipped
folder and is where the archive folders will be situated.}
}
\value{
It will create a folder structure (tile > date) for processing of
 the Sentinel download and return the file path the end location.
}
\description{
\code{s_mk_fold} will construct a nested folder structure of the format,
tile > date. The details are taken from a downloaded Sentinel 2 data zipped
file name and the folders are used to house the zipped file for all future
processing actions.
}
