% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auscophub_functions.R
\name{s_organise}
\alias{s_organise}
\title{Internal function for unzipping and compiling Sentinel 2 data.}
\usage{
s_organise(dzip = dzip, path = path, res = res, ext = ext)
}
\arguments{
\item{dzip}{is a character filepath for one zipped Sentinel download which
is passed from the parent function.}

\item{path}{is a character filepath to the location of the Sentinel-2 zipped
file and is where the archive folders will be situated. This will be
passed from the parent function.}

\item{res}{is a character representation of the required pixel resolution and
can be ONE of "10", "20", "60" or "all" to produce all three possible. This
will be passed from the parent function.}

\item{ext}{is a character representation of the desired file extension for
the output raster stack/s. Intended to be either "tif" or "img" but would
handle any format that GDAL supports. This will be passed from the parent
function.}
}
\value{
It will create intended raster stack/s outputs and leave a clean
directory minus everything bar the original data download and a metadata xml.
}
\description{
\code{s_organise} will unzip a Sentinel-2 downloaded file and compile the
jp2 spectral band data into a multi band raster. It can compile specific
spectral bands based on a provided pixel resolution (10m, 20m or 60m) or can
compile for all three. On completion it cleans up by leaving only the
downloaded zip file, the raster stack/s and an xml with original metadata.
}
