% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auscophub_functions.R
\name{sent_dirs}
\alias{sent_dirs}
\title{Create archive folder directory for Sentinel-2 data downloads}
\usage{
sent_dirs()
}
\value{
This function will create a directory for the tile and the date of
 imagery (all extracted from the zip file name) if they do not already exist.
}
\description{
\code{sent_dirs} is a function that will create a directory tree which will
 be the final resting place for extracted files from the downloaded zip file.
 Specifically it will create for each downloaded zip file a tile directory
 (if it doesn't already exist) and within that a folder named for the date of
 image capture. It is written to work from the location of the zip files.
}
\examples{
\dontrun{
sent_dirs()}
}
