% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auscophub_functions.R
\name{sent_down}
\alias{sent_down}
\title{Download Sentinel-2 data from a list}
\usage{
sent_down(wdir, ddir, aoi = "aoi_list.txt")
}
\arguments{
\item{wdir}{A character string representing the file path to your working
directory which will be the location of the text file of data to download.
File path should be of the type Z:/DOCUMENTATION/..." or double up the
backslashes if on Windows}

\item{ddir}{A character string representing the file path to your download
directory which will be where you want the data to download to. File path
should be of the type Z:/DOCUMENTATION/..." or double up the backslashes if
on Windows.}

\item{aoi}{A character string of the name of the text file that contains the
addresses of Sentinel data to download. It must include the file extension.
The default is "aoi_list.txt"}
}
\value{
This function downloads zipped files of Sentinel-2 data from the
Australian Regional Copernicus Hub to a destination of your choice.
}
\description{
\code{sent_down} is a function to download Sentinel-2 data from a list
generated from a search of the Australian Regional Copernicus Hub. The list
is generated by using the auscophub_searchServer.py python script. See
\href{http://auscophub.readthedocs.io/en/latest/}{here} or the package
vignette for details.
}
\examples{
\dontrun{
sent_down(wdir = "Z:/blah/working", ddir "Z:/blah/downloads", aoi = "test.txt")}

}
