% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auscophub_functions.R
\name{sent_sort}
\alias{sent_sort}
\title{Extract jp2 band data from Sentinel-2 downloaded zip files into local archive}
\usage{
sent_sort(topdir = "Y:/sentinel", ddir = "Y:/sentinel/zdownloads")
}
\arguments{
\item{topdir}{A character string representing the top level in the local
archive. The default is the RS Section's sentinel archive "Y:/sentinel".}

\item{ddir}{A character string representing the download folder in the local
archive. The default is the RS Section's sentinel archive
"Y:/sentinel/zdownloads.}
}
\value{
This function will manage the downloaded Sentinel-2 data zip files
into a folder structure similar to the RS section USGS Landsat archive.
}
\description{
\code{sent_sort} will unzip downloaded Sentinel-2 data to a temp file,
extract just the jp2 formatted bands, create a local archive
(sentinel/TILE/YYYYMMDD) and move the products and original zip file to the
new archive.
}
\note{
This function will:
\itemize{
  \item Create a directory tree which includes a folder for the tile name (if
  it doesn't exist) and within that, a folder for the data overpass date.
  \item Extract all jp2 band data to a temporary file and then copy to the
  appropriate archive folder.
  \item Move the original zip file to the appropriate archive folder}
}
\examples{
\dontrun{
sent_sort()}

}
