% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auscophub_functions.R
\name{sentinel_extract}
\alias{sentinel_extract}
\title{Compile Sentinel-2 data into multiband rasters}
\usage{
sentinel_extract(path = path, res = "10", ext = "tif")
}
\arguments{
\item{path}{is a character filepath to the location of the Sentinel-2 zipped
file/s. Note that this is also where the processed data will be saved.}

\item{res}{is a character representation of the required pixel resolution and
can be ONE of "10", "20", "60" or "all" to produce all three possible. Default
is for the 10 metre optical bands.}

\item{ext}{is a character representation of the desired file extension for
the output raster stack/s. Intended to be either "tif" or "img" but would
handle any format that GDAL supports. Default is "tif".}
}
\value{
all zipped files found will be processed to produce compiled raster
stacks in format of choice. Processed data will be stored in appropriate
folders (tile > date). The compiled data, a metadata xml file and
the zipped data for the original download will all be housed at this location.
}
\description{
\code{sentinel_extract} will unzip all Sentinel-2 downloaded files and
compile the jp2 spectral band data into multi band rasters. It can compile
groups of spectral bands based on  pixel resolution (10m, 20m or
60m) or can compile for all three. On completion it cleans up by leaving only
the downloaded zip file, the raster stack/s and an xml with original metadata
in an archivable folder structure (tile > date). It is intended that the
folder/s and their contents be moved to a final location which could be
dependent on processing level of the downloaded data (e.g. L1C or L2A).
}
\examples{
\dontrun{
sentinel_extract(path = "C:/downloads", res = "60", ext = "img")}

}
